First attempt at parallelization

+++ src/camera.rs
@@ -4,8 +4,8 @@ use vec3::Vec3;
 pub struct Camera {
     pub origin: Vec3,
     pub lower_left: Vec3,
-    pub width: f32,
-    pub height: f32,
+    pub width: usize,
+    pub height: usize,
 }
 
 impl Camera {
@@ -13,8 +13,8 @@ impl Camera {
         return Ray {
             origin: self.origin,
             direction: self.lower_left +
-                Vec3::new2d(u * self.lower_left.x.abs() * 2.0 / self.width,
-                            v * self.lower_left.y.abs() * 2.0 / self.height)
+                Vec3::new2d(u * self.lower_left.x.abs() * 2.0 / self.width as f32,
+                            v * self.lower_left.y.abs() * 2.0 / self.height as f32)
                 - self.origin
         }
     }
